requirements.txt
djangorestframework==3.6.3
Django==1.11
